+++ CHANGELOG.md
@@ -15,6 +15,7 @@ nav_order: 1
 - Removes `receiver_ingesting_remote_write_uri` and `store_uri` Thanos connection info fields
 - Adds `stringtype` to `flink_external_postgresql_user_config` service integration
 - Fix `terraform import` for services with additional disk space or read replica service integration
+- Run client-side validation for `aiven_kafka_schema` AVRO type schema
 
 ## [4.26.0] - 2024-09-25